+++ src/FSAL/FSAL_POSIX/DBExt/MYSQL/posixdb_lock.c
@@ -31,7 +31,7 @@ fsal_posixdb_status_t fsal_posixdb_lockHandleForUpdate(fsal_posixdb_conn * p_con
 
   snprintf(query, 2048, "SELECT handleid, handlets, nlink, ctime, ftype "
            "FROM Handle WHERE deviceid=%llu AND inode=%llu "
-           "FOR UPDATE", p_info->devid, p_info->inode);
+           "FOR UPDATE", (unsigned long long)p_info->devid, (unsigned long long)p_info->inode);
 
   st = db_exec_sql(p_conn, query, &res);
   if(FSAL_POSIXDB_IS_ERROR(st))
